+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-13 13:08:55
-// Commit: feat(perf): implement responsive perf view component
+// Updated at: 2026-05-13 23:25:54
+// Commit: perf(perf): apply RepaintBoundary on animated perf widgets
 
 class perfComponentState {
-  final String commitRef = '1100';
-  final String lastMessage = 'feat(perf): implement responsive perf view component';
-  final DateTime timestamp = DateTime.parse('2026-05-13T13:08:55+05:00');
+  final String commitRef = '1127';
+  final String lastMessage = 'perf(perf): apply RepaintBoundary on animated perf widgets';
+  final DateTime timestamp = DateTime.parse('2026-05-13T23:25:54+05:00');
 }